backend/requirements.txt
# ChainWatch Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
web3>=6.15.0
sqlalchemy>=2.0.0
scikit-learn>=1.4.0
networkx>=3.2.0
numpy>=1.26.0
requests>=2.31.0
pydantic>=2.6.0
aiohttp>=3.9.0
